==> notebook/4_User_Item-basedCF.ipynb <==
<a class="anchor" id="section4"></a>

## Section 4: User-based CF & Item-based CF Recommenders

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

import pandas as pd
import numpy as np

df_train = pd.read_csv('data/split/train_user_artists.csv')

df_train.head(20)

user_id = df_train.userID.unique().tolist()
# Make dictionary ('original_userID': userID)
user_id_dict = dict(zip(user_id, np.arange(1, len(user_id) + 1)))

artist_id = df_train.artistID.unique().tolist()
# Make dictionary ('original_artistID': artistID)
artist_id_dict = dict(zip(artist_id, np.arange(1, len(artist_id) + 1)))

# Reindexed dataframe
remap_id_df_train = df_train.replace({'userID': user_id_dict})
remap_id_df_train = remap_id_df_train.replace({'artistID': artist_id_dict})
remap_id_df_train.head(20)

unique_artist = remap_id_df_train.drop_duplicates(subset ="artistID")
unique_artist = unique_artist[['artistID', 'name']]
unique_artist

df_train = remap_id_df_train[['artistID', 'userID', 'weight']]
df_train.head()

df_train_pivot = df_train.pivot(index="artistID", columns="userID", values="weight").fillna(0)
df_train_pivot.shape  # 12866 artist, 1892 user

df_train_pivot

def create_pivot_table(pd_df):
    data = pd_df.values
    #print(data.shape)
    rows, row_pos = np.unique(data[:, 0], return_inverse=True)
    cols, col_pos = np.unique(data[:, 1], return_inverse=True)
    pivot_matrix = np.zeros((len(rows), len(cols)), dtype=data.dtype)
    pivot_matrix[row_pos, col_pos] = data[:, 2]
    print(pivot_matrix.shape)
    return pivot_matrix

weights = create_pivot_table(df_train)

<a class="anchor" id="section41"></a>

### Section 4.1: Calculate similarity matrix for user & item

* **User Similarity Matrix (#user, #user)**: what is the similarity score between users
* **Item Similarity Matrix (#artist, #artist)**: what is the similarity score between artists

from sklearn.metrics.pairwise import pairwise_distances

# User Similarity Matrix
user_correlation = 1 - pairwise_distances(weights.T, metric='cosine')
user_correlation[np.isnan(user_correlation)] = 0      #set the NaN values in the matrix to O
print('Shape of User Similarity Matrix:', user_correlation.shape)
# Item Similarity Matrix
item_correlation = 1 - pairwise_distances(weights, metric='cosine')
item_correlation[np.isnan(item_correlation)] = 0
print('Shape of Item Similarity Matrix:', item_correlation.shape)

<a class="anchor" id="section42"></a>

### Section 4.2: Predict based on item-based or user-based methods
* User-based:
$$\hat{r}_{ui} = \mu_u + \frac{ \sum\limits_{v \in N^k_i(u)}
\text{sim}(u, v) \cdot (r_{vi} - \mu_v)} {\sum\limits_{v \in
N^k_i(u)} \text{sim}(u, v)}$$
* Item-based:
$$\hat{r}_{ui} = \mu_i + \frac{ \sum\limits_{j \in N^k_u(i)}
\text{sim}(i, j) \cdot (r_{uj} - \mu_j)} {\sum\limits_{j \in
N^k_u(i)} \text{sim}(i, j)}$$

def predict(ratings, similarity, type='user'):
    if type == 'user':
        mean_user_rating = ratings.mean(axis=1)   #axis=0: apply vertical(row)/axis=1: apply horizontial(row)
        # Use np.newaxis so that mean_user_rating has same format as ratings
        ratings_diff = (ratings - mean_user_rating[:, np.newaxis])
        pred = mean_user_rating[:, np.newaxis] + similarity.dot(ratings_diff) / np.array([np.abs(similarity).sum(axis=1)]).T
        #pred = 0 + similarity.dot(ratings_diff) / np.array([np.abs(similarity).sum(axis=1)]).T
    elif type == 'item':
        mean_item_rating = ratings.mean(axis=0)
        ratings_diff = (ratings - mean_item_rating[np.newaxis, :])
        #pred = 0 + ratings_diff.dot(similarity) / np.array([np.abs(similarity).sum(axis=1)])
        pred = mean_item_rating[np.newaxis, :] + ratings_diff.dot(similarity) / np.array([np.abs(similarity).sum(axis=1)])
    return pred.clip(min=0)

user_prediction = predict(weights.T, user_correlation, type='user')
item_prediction = predict(weights.T, item_correlation, type='item')

<a class="anchor" id="section43"></a>

### Section 4.3 Recommend artists based on user preference

user_pred_df = pd.DataFrame(user_prediction, columns =list(unique_artist.artistID))
item_pred_df = pd.DataFrame(item_prediction, columns = list(unique_artist.artistID))

#adopted from Web Mining exercise 1 ( movie represents artists here)
def recommend_movies(pred_df, userID, movies, original_ratings, num_recommendations, method_name):
    # Get and sort the user's predictions
    user_row_number = userID - 1 # User ID starts at 1, not 0
    sorted_user_predictions = pred_df.iloc[user_row_number].sort_values(ascending=False) # User ID starts at 1
    
    # Get the user's data and merge in the movie information.
    user_data = original_ratings[original_ratings.userID == (userID)]
    user_full = (user_data.merge(movies, how = 'left', left_on = 'artistID', right_on = 'artistID').
                     sort_values(['weight'], ascending=False)
                 )
    
    # Recommend the highest predicted rating movies that the user hasn't seen yet.
    recommendations = (movies[~movies['artistID'].isin(user_full['artistID'])].
         merge(pd.DataFrame(sorted_user_predictions).rename_axis('artistID').reset_index(), how = 'left',
               left_on = 'artistID',
               right_on = 'artistID').
         rename(columns = {user_row_number: method_name}).
         sort_values(method_name, ascending = False).iloc[:num_recommendations, :]
         )
    recommendations.drop(recommendations[recommendations[method_name] == 0.0].index, inplace=True)

    return user_full, recommendations

# Result example before mapping back userID/artistID
print('Collaborative Filtering RS (user-based)')
user_full, recommendations = recommend_movies(user_pred_df, 1, unique_artist, df_train, 10,'CF_user_pred_rating')
recommendations

user_based_result = recommendations.iloc[0:0]
user_based_result.insert(0, "userID", None)

# Result example before mapping back userID/artistID
print('Collaborative Filtering RS (item-based)')
user_full, recommendations = recommend_movies(item_pred_df, 2, unique_artist, df_train, 10, 'CF_item_pred_rating')
recommendations

item_based_result = recommendations.iloc[0:0]
item_based_result.insert(0, "userID", None)

# inverse dictionary for mapping back userID 
inv_user_id_dict = {v: k for k, v in user_id_dict.items()}
inv_artist_id_dict = {v: k for k, v in artist_id_dict.items()}

# Merge User-based result 
no_recommendation_user = []
for i in np.arange(1, len(df_train_pivot.columns) + 1):
    try:
        _, recommendations = recommend_movies(user_pred_df, i, unique_artist, df_train, 10, 'CF_user_pred_rating')
        recommendations['userID'] = i
        user_based_result = user_based_result.append(recommendations, ignore_index=True)
    except:
        print("User {} doesn't exist recommendation".format(i))
        no_recommendation_user.append(i)

# Map back userID
user_mapback = user_based_result.replace({'userID': inv_user_id_dict})
user_mapback = user_mapback.replace({'artistID': inv_artist_id_dict})
user_based_result = user_mapback.sort_values(by=['userID','CF_user_pred_rating'], ascending = [True, False])
user_based_result.head(20)

user_based_result.to_csv('data/result/user_based_result.csv', index=False)

# Merge item-based result
no_recommendation_item = []
for i in np.arange(1, len(df_train_pivot.columns) + 1):
    try:
        _, recommendations = recommend_movies(item_pred_df, i, unique_artist, df_train, 10,'CF_item_pred_rating')
        recommendations['userID'] = i
        item_based_result = item_based_result.append(recommendations, ignore_index=True)
    except:
        print("User {} doesn't exist recommendation".format(i))
        no_recommendation_item.append(i)

# Map back userID
item_mapback = item_based_result.replace({'userID': inv_user_id_dict})
item_mapback = item_mapback.replace({'artistID': inv_artist_id_dict})
item_based_result = item_mapback.sort_values(by=['userID','CF_item_pred_rating'], ascending = [True, False])
item_based_result.head(20)

item_based_result.to_csv('data/result/item_based_result.csv', index=False)
