==> notebook/6_Evaluation.ipynb <==
<a class="anchor" id="section6"></a>

## Section 6 Evaluation

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

import pandas as pd
import numpy as np
import math

df_test = pd.read_csv('data/split/test_user_artists.csv')
user_based_result = pd.read_csv('data/result/user_based_result.csv')
item_based_result = pd.read_csv('data/result/item_based_result.csv')
personalrank_tags = pd.read_csv('data/result/withtag.csv')
personalrank_wo_tags = pd.read_csv('data/result/without_tag.csv')
content_based_result = pd.read_csv('data/result/content_based_result.csv')
df_train = pd.read_csv('data/split/train_user_artists.csv')

merged_genre = pd.read_csv('data/external/merged_genre.csv')

# Prepare the novelty for each artist
df_train = df_train[['name', 'userID',	'artistID',	'weight']]
artist_sum = df_train.groupby(by='artistID').count().drop(['name','weight'], axis=1).rename(columns={"userID": "user_count"})
artist_sum['novelty'] = artist_sum['user_count'].apply(lambda x: -1*math.log2(x/len(df_train.userID.unique()))/(-1*math.log2(1/len(df_train.userID.unique()))))
artist_sum.head()

df_test = df_test[['userID','artistID', 'name',	'weight']]
print('{} unique user in test set'.format(len(df_test.userID.unique())))
df_test.head()

# Calculate Recall@10 and Precision@10 
def evaluation_metric(df_test, result):

    score = pd.DataFrame(columns={'userID','recall@10','precision@10'})
    fail = []
    user_list = result.userID.unique()
    for i in user_list:
        pred = result.loc[result['userID'] == i, 'artistID'].tolist()
        true = df_test.loc[df_test['userID'] == i].sort_values(by=['weight'],ascending = False)
        if len(true)>= 10:
            true = true.artistID.tolist()
            recall = len(set(true).intersection(pred))/len(true)
            precision = len(set(true).intersection(pred))/10
            score = score.append(pd.DataFrame(data={'userID': [i], 'recall@10': [recall], 'precision@10': [precision]}))
        else:
            fail.append(i)  
          
    return score, fail


def evaluation_report(result):
    print('{} unique user in file'.format(len(result.userID.unique())))
    score, fail = evaluation_metric(df_test, result)
    score = score[['userID','recall@10','precision@10']]
    print('{} users have less than 10 artists in test set and fail to do evaluation'.format(len(fail)))
    not_zero = score.dropna()
    not_zero = not_zero.loc[not_zero['recall@10']!= 0.0]
    print('Total number of user able to do evaluation with k=10: {}'.format(len(score)))
    print('There are {} users who has non-zero hits'.format(len(not_zero)))
    print('The average of recall@10: {}'.format(score['recall@10'].mean()))
    print('The average of precision@10: {}'.format(score['precision@10'].mean()))    

#Calculate diversity
def diversity_calculation(result, merged_genre):
    diversity = pd.DataFrame(columns={'userID','diversity'})
    user_list = result.userID.unique()
    for i in user_list:
        pred = result.loc[result['userID'] == i, 'artistID'].tolist()
        genre_hit = merged_genre[merged_genre['id'].isin(pred)]
        merged_list = []
        for i in genre_hit.index:
            for item in genre_hit.genre_merge[i].strip("[]").replace("'","").replace("' ","").split(","):
                merged_list.append(item.strip())
        diversity = diversity.append(pd.DataFrame(data={'userID': [i], 'diversity': [len(set(merged_list))]}))
    return diversity   

def diversity_report(result):
    diversity = diversity_calculation(result, merged_genre)
    print('The average of diversity: {}'.format(diversity['diversity'].mean()))
    print('The median of diversity: {}'.format(diversity['diversity'].median()))                                            

#Calculate novelty and coverage
def novelty_report(result, artist_sum):
    table = result.merge(artist_sum, left_on='artistID', right_on='artistID')
    print('Novelty: {}'.format(table['novelty'].mean()))

def coverage_report(result, df_train):
    coverage = len(result.artistID.unique())/len(df_train.artistID.unique())
    print('Coverage: {}'.format(coverage))

<a class="anchor" id="section61"></a>
### Section 6.1 User-based metrics

user_based_result = user_based_result[user_based_result['userID'].isin(df_test.userID.unique().tolist())]

evaluation_report(user_based_result)

diversity_report(user_based_result)

novelty_report(user_based_result, artist_sum)

coverage_report(user_based_result,df_train)

<a class="anchor" id="section62"></a>
### Section 6.2 Item-based metrics

item_based_result = item_based_result[item_based_result['userID'].isin(df_test.userID.unique().tolist())]

evaluation_report(item_based_result)

diversity_report(item_based_result)

novelty_report(item_based_result, artist_sum)

coverage_report(item_based_result,df_train)

<a class="anchor" id="section63"></a>
### Section 6.3 Content-based 

content_based_result = content_based_result[['userID','artistID']]
content_based_result = content_based_result[content_based_result['userID'].isin(df_test.userID.unique().tolist())]

evaluation_report(content_based_result)

diversity_report(content_based_result)

novelty_report(content_based_result, artist_sum)

coverage_report(content_based_result,df_train)

<a class="anchor" id="section64"></a>
### Section 6.4 PersonalRank without tags

personalrank_wo_tags = personalrank_wo_tags.rename(columns={"recom_artist": "artistID"})
personalrank_wo_tags = personalrank_wo_tags[['userID','artistID']]

evaluation_report(personalrank_wo_tags)

diversity_report(personalrank_wo_tags)

novelty_report(personalrank_wo_tags, artist_sum)

coverage_report(personalrank_wo_tags, df_train)

<a class="anchor" id="section65"></a>
### Section 6.5 PersonalRank with tags

personalrank_tags = personalrank_tags[['userID', "recom_artist"]]
personalrank_tags = personalrank_tags.rename(columns={"recom_artist": "artistID"})

evaluation_report(personalrank_tags)

diversity_report(personalrank_tags)

novelty_report(personalrank_tags, artist_sum)

coverage_report(personalrank_tags, df_train)
